requirements.txt
# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
pyarrow>=14.0

# Data sources
yfinance>=0.2
# dukascopy-python (optional — tick-level FX); install separately if needed
# ccxt (optional — crypto)

# Visualization
matplotlib>=3.7
mplfinance>=0.12
plotly>=5.17

# Utils
pyyaml>=6.0
python-dateutil>=2.8
pytz>=2023.3
tqdm>=4.66

# Testing
pytest>=7.4
pytest-cov>=4.1
hypothesis>=6.90
